tldr: reordering is not swapping, duh!
this turned into a whole thing... But the basic problem is that item descriptions do not have a list order. Instead of changing the underlying database structure, the workaround is to point the description's id at new data. In this way, we can reorder descriptions without reordering anything in the database. This is bad, I know. Because this will break anything that relies on any specific description id.

BUT, I know that nothing relies on description id's, and they're always deleted permenently, so it's is ok for now. On top of that, it will be easy to geet this pattern when I finally change the way the database manages item descriptions.